remove http state (#27)
