+++ src/appState/appState.ts
@@ -1,3 +1,5 @@
 export * from './accountState';
 export * from './providerState';
 export * from './tokenState';
+export * from './util';
+export * from './updateStateModel';
